+++ ambition_utils/rrule/tests/model_tests.py
@@ -1,6 +1,7 @@
 import datetime
 import fleming
 
+from fleming import fleming
 import pytz
 from dateutil import rrule, parser
 from django.test import TestCase
@@ -649,7 +650,7 @@ class RRuleTest(TestCase):
 
     def test_generate_dates(self):
         """
-        Test a monthly first day of month rule to catch case of converting tz back using the generate_dates method
+        Assert generate_dates returns the same values as get_dates.
         """
         params = {
             'freq': rrule.MONTHLY,
@@ -664,23 +665,122 @@ class RRuleTest(TestCase):
             time_zone=pytz.timezone('US/Eastern'),
             occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
         )
-        next_dates = rule.generate_dates()
+        
+        self.assertEqual(
+            rule.get_dates(),
+            rule.generate_dates()
+        )
+        
+    def test_get_dates(self):
+        """
+        Test a monthly 1st day of month rule to catch case of converting tz back using the get_dates method
+        """
+        params = {
+            'freq': rrule.MONTHLY,
+            'interval': 1,
+            'dtstart': datetime.datetime(2016, 12, 31),
+            'bymonthday': 1,
+            'until': datetime.datetime(2017, 4, 30),
+        }
+
+        rule = RRule(
+            rrule_params=params,
+            time_zone=pytz.timezone('US/Eastern'),
+            occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
+        )
+        next_dates = rule.get_dates()
 
         # Check a few dates
         self.assertEqual(len(next_dates), 4)
-        self.assertEqual(next_dates[0], datetime.datetime(2017, 2, 1, 3))
-        self.assertEqual(next_dates[1], datetime.datetime(2017, 3, 1, 3))
-        self.assertEqual(next_dates[2], datetime.datetime(2017, 4, 1, 2))
-        self.assertEqual(next_dates[3], datetime.datetime(2017, 5, 1, 2))
+
+        self.assertEqual(next_dates[0], datetime.datetime(2017, 1, 1, 5))
+        self.assertEqual(next_dates[1], datetime.datetime(2017, 2, 1, 5))
+        self.assertEqual(next_dates[2], datetime.datetime(2017, 3, 1, 5))
+        self.assertEqual(next_dates[3], datetime.datetime(2017, 4, 1, 4))  # DST change for US/Eastern
 
         # Run pre save again to make sure it doesn't mess up params
         rule.pre_save_hooks()
 
         # Get next dates to compare against
-        more_next_dates = rule.generate_dates()
+        more_next_dates = rule.get_dates()
         self.assertEqual(next_dates, more_next_dates)
+        
+    def test_get_dates_with_start_date(self):
+        """
+        Test a date generation with a start date.
+        """
+        params = {
+            'freq': rrule.MONTHLY,
+            'interval': 1,
+            'dtstart': datetime.datetime(2016, 12, 31),
+            'bymonthday': 1,
+        }
+
+        rule = RRule(
+            rrule_params=params,
+            time_zone=pytz.timezone('US/Eastern'),
+            occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
+        )
+        next_dates = rule.get_dates(num_dates=10, start_date=datetime.datetime(2018, 1, 1))
+
+        # Check a few dates
+        self.assertEqual(len(next_dates), 10)
+        self.assertEqual(next_dates[0], datetime.datetime(2018, 1, 1, 5))
+        self.assertEqual(next_dates[1], datetime.datetime(2018, 2, 1, 5))
+        self.assertEqual(next_dates[2], datetime.datetime(2018, 3, 1, 5))
+        self.assertEqual(next_dates[3], datetime.datetime(2018, 4, 1, 4))  # DST change for US/Eastern
+        self.assertEqual(next_dates[-1], datetime.datetime(2018, 10, 1, 4))
+
+    def test_get_dates_num_dates_greater(self):
+        """
+        Test a date generation with a start date and end date that will yield fewer dates than num_dates.
+        Daily from 1/1 to 1/10 is 10. Request default of 20 but only after 1/5 which should yield 6 dates, inclusive.
+        """
+        rule = RRule(
+            rrule_params={
+                'freq': rrule.DAILY,
+                'interval': 1,
+                'dtstart': datetime.datetime(2017, 1, 1),
+                'until': datetime.datetime(2017, 1, 10),
+            },
+            time_zone=pytz.timezone('US/Eastern'),
+            occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
+        )
+        next_dates = rule.get_dates(
+            num_dates=20,
+            start_date=datetime.datetime(2017, 1, 5)
+        )
+
+        self.assertEqual(len(next_dates), 6)
+        self.assertEqual(next_dates[0], datetime.datetime(2017, 1, 5, 5))
+        self.assertEqual(next_dates[1], datetime.datetime(2017, 1, 6, 5))
+        self.assertEqual(next_dates[2], datetime.datetime(2017, 1, 7, 5))
+        self.assertEqual(next_dates[3], datetime.datetime(2017, 1, 8, 5))
+        self.assertEqual(next_dates[4], datetime.datetime(2017, 1, 9, 5))
+        self.assertEqual(next_dates[5], datetime.datetime(2017, 1, 10, 5))
 
     def test_generate_dates_from_params(self):
+        """
+        Assert generate_dates_from_params returns the same values as get_dates_from_params.
+        """
+        params = {
+            'rrule_params': {
+                'freq': rrule.MONTHLY,
+                'interval': 1,
+                'dtstart': datetime.datetime(2017, 1, 1, 22),
+                'bymonthday': -1,
+                'until': datetime.datetime(2017, 5, 1, 22),
+            },
+            'time_zone': pytz.timezone('US/Eastern'),
+            'num_dates': 3,
+        }
+
+        self.assertEqual(
+            RRule.get_dates_from_params(**params),
+            RRule.generate_dates_from_params(**params)
+        )
+
+    def test_get_dates_from_params(self):
         """
         Tests the class method wrapper
         """
@@ -697,19 +797,23 @@ class RRuleTest(TestCase):
             time_zone=pytz.timezone('US/Eastern'),
             occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
         )
-        next_dates = rule.generate_dates(num_dates=3)
+        next_dates = rule.get_dates(
+            num_dates=3, 
+            start_date=datetime.datetime(2017, 2, 2, 22)
+        )
 
-        next_dates_from_params = RRule.generate_dates_from_params(
+        next_dates_from_params = RRule.get_dates_from_params(
             rrule_params=params,
             time_zone=pytz.timezone('US/Eastern'),
             num_dates=3,
+            start_date=datetime.datetime(2017, 2, 2, 22)
         )
 
         self.assertEqual(next_dates, next_dates_from_params)
 
-    def test_model_different_time_zone_end_of_month_generate_dates(self):
+    def test_model_different_time_zone_end_of_month_get_dates(self):
         """
-        Test a monthly first day of month rule to catch case of converting tz back using the generate_dates method
+        Test a monthly first day of month rule to catch case of converting tz back using the get_dates method
         """
         params = {
             'freq': rrule.MONTHLY,
@@ -723,7 +827,7 @@ class RRuleTest(TestCase):
             time_zone=pytz.timezone('US/Eastern'),
             occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
         )
-        next_dates = rule.generate_dates()
+        next_dates = rule.get_dates()
         self.assertEqual(next_dates[0], datetime.datetime(2017, 2, 1, 3))
         self.assertEqual(next_dates[1], datetime.datetime(2017, 3, 1, 3))
         self.assertEqual(next_dates[2], datetime.datetime(2017, 4, 1, 2))
@@ -1044,7 +1148,7 @@ class RRuleTest(TestCase):
         self.assertEqual(rule.rrule_params, clone.rrule_params)
 
         # Assert the generated dates are equal.
-        self.assertEqual(rule.generate_dates(num_dates=4), clone.generate_dates(num_dates=4))
+        self.assertEqual(rule.get_dates(num_dates=4), clone.get_dates(num_dates=4))
 
     @freeze_time('6-15-2022')
     def test_weekly_clone_with_offset(self):
@@ -1053,7 +1157,7 @@ class RRuleTest(TestCase):
         rule = RRule.objects.create(
             rrule_params={
                 'freq': rrule.WEEKLY,
-                'dtstart': datetime.datetime(2022, 6, 22),
+                'dtstart': datetime.datetime(2022, 6, 21),  # Tuesday
                 'byweekday': [0, 2, 4],
             },
             occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
@@ -1078,37 +1182,34 @@ class RRuleTest(TestCase):
 
         # Assert the generated dates are as expected.
         self.assertEqual(
-            rule.generate_dates(num_dates=4),
+            rule.get_dates(num_dates=4),
             [
                 datetime.datetime(2022, 6, 22),  # Wednesday
                 datetime.datetime(2022, 6, 24),  # Friday
                 datetime.datetime(2022, 6, 27),  # Monday
                 datetime.datetime(2022, 6, 29),  # Wednesday
-                datetime.datetime(2022, 7, 1),   # Friday
             ]
         )
 
         # Two days after each date in the regular series.
         self.assertEqual(
-            future_clone.generate_dates(num_dates=4),
+            future_clone.get_dates(num_dates=4),
             [
                 datetime.datetime(2022, 6, 24),  # Friday
                 datetime.datetime(2022, 6, 26),  # Sunday
                 datetime.datetime(2022, 6, 29),  # Wednesday
                 datetime.datetime(2022, 7, 1),   # Friday
-                datetime.datetime(2022, 7, 3),   # Sunday
             ]
         )
 
         # Two days before each date in the regular series.
         self.assertEqual(
-            past_clone.generate_dates(num_dates=4),
+            past_clone.get_dates(num_dates=4),
             [
                 datetime.datetime(2022, 6, 20),  # Monday
                 datetime.datetime(2022, 6, 22),  # Wednesday
                 datetime.datetime(2022, 6, 25),  # Saturday
                 datetime.datetime(2022, 6, 27),  # Monday
-                datetime.datetime(2022, 6, 29),  # Wednesday
             ]
         )
 
@@ -1175,7 +1276,7 @@ class RRuleTest(TestCase):
 
         # Assert the generated dates are as expected for the original.
         self.assertEqual(
-            rule.generate_dates(),
+            rule.get_dates(),
             [
                 datetime.datetime(2022, 10, 15),
                 datetime.datetime(2022, 10, 16),
@@ -1185,7 +1286,7 @@ class RRuleTest(TestCase):
 
         # One day after each date in the regular series.
         self.assertEqual(
-            future_clone.generate_dates(),
+            future_clone.get_dates(),
             [
                 datetime.datetime(2022, 10, 16),
                 datetime.datetime(2022, 10, 17),
@@ -1195,7 +1296,7 @@ class RRuleTest(TestCase):
 
         # One day before each date in the regular series.
         self.assertEqual(
-            past_clone.generate_dates(),
+            past_clone.get_dates(),
             [
                 datetime.datetime(2022, 10, 14),
                 datetime.datetime(2022, 10, 15),